Fix compiling with Xcode 15.
In the iOS 18 SDK, `SKNode` conforms to `UICoordinateSpace`, which means that it now has `bounds` property. Our implementation would then need to override it. But if we add `override`, then the code fails to compile on Xcode 15 with the iOS 17 SDK, because in that platform, there's nothing to override.

We're actually not trying to override anything; we just wanted a convenience property for accessing the view's bounds. Given that we're running into a naming collision, the simplest approach is to just change the name of this property.